revisi

+++ database/migrations/2020_11_05_233100_symptom_check.php
@@ -15,8 +15,8 @@ class SymptomCheck extends Migration
     {
         Schema::create('symptom_checks', function (Blueprint $table) {
             $table->id();
-            $table->integer("day_id");
-            $table->integer("symptom_id");
+            $table->unsignedBigInteger("day_id");
+            $table->unsignedInteger("symptom_id");
             $table->string("data")->nullable();
             $table->unsignedTinyInteger("status");
             $table->timestamps();
